web-animations/core/animation-timeline.html

<!doctype html>
<title>Web Animations API: AnimationTimeline tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="support/iframe.html" width="10" height="10" id="iframe"></iframe>
<script>

test(function() {
  assert_equals(typeof document.timeline, 'object',
    'document.timeline is not undefined');
  assert_not_equals(document.timeline, null,
    'document.timeline is not null');
  assert_equals(document.timeline, document.timeline,
    'document.timeline returns the same object every time');
  var iframe = document.getElementById('iframe');
  assert_not_equals(document.timeline, iframe.contentDocument.timeline,
    'document.timeline returns a different object for each document');
},
'document.timeline identity tests',
{
  help:   'http://dev.w3.org/fxtf/web-animations/#the-document-timeline',
  assert: [ 'Each document has a timeline called the document timeline' ],
  author: 'Brian Birtles'
});

test(function() {
  assert_equals(typeof document.timeline.currentTime, 'number',
    'document.timeline.currentTime returns a number');
  assert_true(document.timeline.currentTime > 0,
    'document.timeline.currentTime is positive');
  if (typeof window.performance.now === 'function') {
    assert_approx_equals(document.timeline.currentTime,
      window.performance.now(), 10000, // 10 seconds drift
      'document.timeline.currentTime is measured from navigationStart');
  }
},
'document.timeline.currentTime value tests',
{
  help: [
    'http://dev.w3.org/fxtf/web-animations/#the-global-clock',
    'http://dev.w3.org/fxtf/web-animations/#the-document-timeline'
  ],
  assert: [
    'The global clock is a source of monotonically increasing time values',
    'The time values of the document timeline are calculated as a fixed ' +
    'offset from the global clock',
    'the zero time corresponds to the navigationStart moment',
    'The time values produced by the global clock represent wall-clock seconds'
  ],
  author: 'Brian Birtles'
});

async_test(function(t) {
  var valueAtStart = document.timeline.currentTime;
  assert_equals(document.timeline.currentTime, valueAtStart,
    'document.timeline.currentTime does not change within a script block');
  window.requestAnimationFrame(function() {
    t.step(function() {
      assert_true(document.timeline.currentTime > valueAtStart,
        'document.timeline.currentTime increases between script blocks');
      t.done();
    });
  });
},
'document.timeline.currentTime liveness tests',
{
  help: 'http://dev.w3.org/fxtf/web-animations/#script-execution-and-live-updates-to-the-model',
  assert: [ 'The value returned by the currentTime attribute of a document timeline will not change within a script block' ],
  author: 'Brian Birtles'
});

</script>
